// src/interfaces/envVars.ts
export interface EnvVars {
  PORT: number;
  STATE: string;
  DB_PASSWORD: string;
  DB_NAME: string;
  DB_HOST: string;
  DB_USERNAME: string;
  JWT_SECRET: string;
  MAIL_HOST: string;
  MAIL_USER: string;
  MAIL_PASSWORD: string;
  MAIL_FROM: string;
  APP_NAME: string;
  FRONTEND_URL: string;
}